refactor: 컴포넌트 및 훅 분리작업
